feat(archon): add Expert Tensor Parallelism (ETP) support for MoE models (#839)
Add ETP support to enable 2D sharding of MoE expert weights across both
expert parallel (EP) and tensor parallel (TP) dimensions.

Key changes:

Parallel dimensions:
- Add `etp` parameter supporting etp=1 (TP borrowed by EP) and etp=tp modes

Expert parallelism:
- Add TensorParallel, ExpertTensorParallel, ReordererSequenceParallel classes
- Extract TokenReorderer from MoE, update apply_moe_ep_tp for strategy selection

Activation checkpointing:
- Fix op-level selective AC not working due to missing op registration

+++ areal/experimental/models/archon/moe/router.py
@@ -192,7 +192,7 @@ class TokenChoiceTopKRouter(nn.Module):
             bins=self.num_experts,
             min=0,
             max=self.num_experts,
-        ).to(torch.int64)
+        )
 
         return top_scores, selected_experts_indices, num_tokens_per_expert
